malloc cache: now using uint64_t instead of size_t

+++ core/bh_base.cpp
@@ -67,7 +67,7 @@ void bh_data_free(bh_base *base) {
     base->data = nullptr;
 }
 
-void bh_data_malloc_stat(size_t &cache_lookup, size_t &cache_misses, size_t &max_memory_usage) {
+void bh_data_malloc_stat(uint64_t &cache_lookup, uint64_t &cache_misses, uint64_t &max_memory_usage) {
     cache_lookup = malloc_cache.getTotalNumLookups();
     cache_misses = malloc_cache.getTotalNumMisses();
     max_memory_usage = malloc_cache.getMaxMemAllocated();
