+++ include/bh_base.hpp
@@ -83,4 +83,4 @@ void bh_data_malloc(bh_base* base);
  */
 void bh_data_free(bh_base* base);
 
-void bh_data_malloc_stat(size_t &cache_lookup, size_t &cache_misses, size_t &max_memory_usage);
+void bh_data_malloc_stat(uint64_t &cache_lookup, uint64_t &cache_misses, uint64_t &max_memory_usage);
